+++ scripts/variables.sh
@@ -4,3 +4,5 @@ REGISTERED_PANE_PREFIX="@-sidebar-registered-pane"
 REGISTERED_SIDEBAR_PREFIX="@-sidebar-is-sidebar"
 
 SUPPORTED_TMUX_VERSION="1.9"
+
+SIDEBAR_DIR="$HOME/.tmux/sidebar"