+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - less does not wrap lines in the sidebar
 - handle issue with refreshing the main pane
 - by default the tree window is now 40 columns wide
+- remember custom sidebar width for each directory
 
 ### v0.3.0, Sep 04, 2014
 - if the pane is too narrow do not show sidebar